requirements.txt
lpips==0.1.4
numpy==1.23.3
opencv_python==4.1.2.30
opencv_python_headless==4.7.0.72
Pillow==9.4.0
Pillow==11.2.1
scipy==1.9.1
timm==0.6.13
torch==1.11.0
torchvision==0.12.0
